requirements.txt
django >= 2.0.1
markdown >= 2.6.11
djangorestframework >= 3.7.7
django-webpack-loader >= 0.5.0